refactor!: separate `Relocation.h`
