+++ CommonLibF4/include/REL/ID.h
@@ -1,0 +1,32 @@
+#pragma once
+
+#include "REL/IDDB.h"
+#include "REL/Module.h"
+
+namespace REL
+{
+	class ID
+	{
+	public:
+		constexpr ID() noexcept = default;
+
+		explicit constexpr ID(std::uint64_t a_id) noexcept :
+			_id(a_id)
+		{}
+
+		constexpr ID& operator=(std::uint64_t a_id) noexcept
+		{
+			_id = a_id;
+			return *this;
+		}
+
+		[[nodiscard]] std::uintptr_t address() const { return base() + offset(); }
+		[[nodiscard]] constexpr std::uint64_t id() const noexcept { return _id; }
+		[[nodiscard]] std::size_t offset() const { return IDDB::get().id2offset(_id); }
+
+	private:
+		[[nodiscard]] static std::uintptr_t base() { return Module::get().base(); }
+
+		std::uint64_t _id{ static_cast<std::uint64_t>(-1) };
+	};
+}
