+++ CommonLibF4/include/REL/REL.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include "REL/ID.h"
+#include "REL/IDDB.h"
+#include "REL/Module.h"
+#include "REL/Offset.h"
+#include "REL/Offset2ID.h"
+#include "REL/Relocation.h"
+#include "REL/Segment.h"
+#include "REL/Version.h"
