+++ CommonLibF4/src/F4SE/API.cpp
@@ -64,7 +64,7 @@ namespace F4SE
 		}
 
 		(void)REL::Module::get();
-		(void)REL::IDDatabase::get();
+		(void)REL::IDDB::get();
 
 		auto& storage = detail::APIStorage::get();
 		const auto& intfc = *a_intfc;
